refactor: extract bundle mesh construction
